fix: convert lex-walkthrough to TS
