+++ .eslint-dictionary.json
@@ -222,6 +222,7 @@
   "kinesis",
   "lang",
   "lerna",
+  "lillies",
   "linux",
   "localhost",
   "localinvoke",
