+++ packages/amplify-cli-core/src/errors/amplify-exception.ts
@@ -175,6 +175,7 @@ export type AmplifyFaultType =
   | 'ConfigurationFault'
   | 'DeploymentFault'
   | 'DeploymentStateUploadFault'
+  | 'FileNotFoundFault'
   | 'LockFileNotFoundFault'
   | 'LockFileParsingFault'
   | 'MockProcessFault'
